refactor: unify patch-cache precompute into one --split script
precompute_patches.py and precompute_val_patches.py were ~90% identical:
the pool scaffolding, memmap write loop, seeding helper, and -- critically
-- the entire shared config block (brains, transform_cfg, offsets,
sampling params) were copy-pasted. That duplicated config was a latent
correctness trap: the train and val caches must use the identical mapping
(the training run applies the val cache's transform.json to the train
cache), enforced only by keeping two blocks in sync.

Replace both with scripts/precompute.py --split {train,val}: one config
block, one scaffolding, branching only on the sampler (TrainDataset mask
= intensity + segmentation vs ValidateDataset intensity-only), the RNG
stream id, and the per-split cache_dir / n_patches. Sampling draws are
bit-identical to the separate scripts at the same seed.

Also apply two behaviors uniformly that previously differed: stamp
transform.json for both splits, and refuse offset calibration for both
(each worker would otherwise calibrate on its own sample, mixing
inconsistent offsets into the cache). Update references in
train_bm4dnet.py and data_handling.py docstrings.
